fix build issue

+++ runtime/src/lib.rs
@@ -274,7 +274,7 @@ impl OperatorProvider<AccountId> for OperatorCollectiveProvider {
 
 parameter_types! {
 	pub const MinimumCount: u32 = 1; // TODO: change this
-	pub const ExpiresIn: u32 = 1000 * 60 * 60 * 24 * 100; // 100days for now TODO: change this
+	pub const ExpiresIn: u64 = 1000 * 60 * 60 * 24 * 100; // 100days for now TODO: change this
 }
 
 impl orml_oracle::Trait for Runtime {
